SQL: Auth db data type cleanup `logs`, `realmcharacters`, `realmlist`, `uptime`  / Storage engine switched to InnoDB Note:
As for the redundant `uptime`.`startstring` being removed, this page might be of interest how to generate time strings from unix timestamps
http://dev.mysql.com/doc/refman/5.1/en/date-and-time-functions.html#function_from-unixtime

+++ src/server/authserver/Realms/RealmList.cpp
@@ -31,7 +31,7 @@ void RealmList::Initialize(uint32 updateInterval)
     UpdateRealms(true);
 }
 
-void RealmList::UpdateRealm(uint32 ID, const std::string& name, const std::string& address, uint32 port, uint8 icon, RealmFlags flag, uint8 timezone, AccountTypes allowedSecurityLevel, float popu, uint32 build)
+void RealmList::UpdateRealm(uint32 ID, const std::string& name, const std::string& address, uint16 port, uint8 icon, RealmFlags flag, uint8 timezone, AccountTypes allowedSecurityLevel, float popu, uint32 build)
 {
     // Create new if not exist or update existed
     Realm& realm = m_realms[name];
@@ -82,7 +82,7 @@ void RealmList::UpdateRealms(bool init)
             uint32 realmId = fields[0].GetUInt32();
             const std::string& name = fields[1].GetString();
             const std::string& address = fields[2].GetString();
-            uint32 port = fields[3].GetUInt32();
+            uint16 port = fields[3].GetUInt16();
             uint8 icon = fields[4].GetUInt8();
             RealmFlags flag = RealmFlags(fields[5].GetUInt8());
             uint8 timezone = fields[6].GetUInt8();
